+++ .ladle/config.mjs
@@ -1,0 +1,3 @@
+export default {
+  stories: "fp-react/src/**/**.stories.{js,jsx,ts,tsx}",
+};