+++ projects/react/fp-button/src/lib/react-fp-button.tsx
@@ -1,0 +1,14 @@
+import styles from "./react-fp-button.module.scss"
+
+/* eslint-disable-next-line */
+export interface ReactFpButtonProps {}
+
+export function ReactFpButton(props: ReactFpButtonProps) {
+  return (
+    <div className={styles["container"]}>
+      <h1>Welcome to ReactFpButton!</h1>
+    </div>
+  )
+}
+
+export default ReactFpButton