+++ react/fp-button/__tests__/FPButton.test.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import { render, screen } from '@testing-library/react';
+import { FPButton } from '../src';
+
+test('renders component successfully', () => {
+  render(<FPButton  />);
+  const element = screen.getByTestId(/test/i);
+  expect(element).toBeInTheDocument();
+});