+++ first-paint/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.16.3](https://github.com/shawn-sandy/fp-kit/compare/@shawnsandy/first-paint@2.16.2...@shawnsandy/first-paint@2.16.3) (2022-07-13)
+
+
+### Features
+
+* ✨ add  style for aria-disabled button ([e7347b3](https://github.com/shawn-sandy/fp-kit/commit/e7347b3f3818914760a12018b0f0c4e98f97ccba))
+
+
+
+
+
+## [2.16.2](https://github.com/shawn-sandy/fp-kit/compare/@shawnsandy/first-paint@2.16.1...@shawnsandy/first-paint@2.16.2) (2022-07-13)
+
+**Note:** Version bump only for package @shawnsandy/first-paint
+
+
+
+
+
+## [2.16.1](https://github.com/shawn-sandy/fp-kit/compare/@shawnsandy/first-paint@2.16.0...@shawnsandy/first-paint@2.16.1) (2022-05-27)
+
+
+### Features
+
+* **beta:** add is:hover:focus to button states ([cd04d31](https://github.com/shawn-sandy/fp-kit/commit/cd04d312bdea26c578cea717e1cae976fad6ec14))
+
+
+
+
+
 # [2.16.0](https://github.com/shawn-sandy/fp-kit/compare/@shawnsandy/first-paint@2.15.3...@shawnsandy/first-paint@2.16.0) (2022-05-14)